공유풀 페이지 API 연동 및 반응형 UI 개선
