+++ src/components/common/InfoTooltip.tsx
@@ -1,0 +1,32 @@
+import { useState } from 'react';
+
+interface InfoTooltipProps {
+  text: string;
+}
+
+export default function InfoTooltip({ text }: InfoTooltipProps) {
+  const [isVisible, setIsVisible] = useState(false);
+
+  return (
+    <div className="relative inline-block">
+      <button
+        type="button"
+        aria-label="정보"
+        className="w-5 h-5 rounded-full bg-[#CCCCCC] flex items-center justify-center text-white cursor-help"
+        style={{ fontSize: '0.75em' }}
+        onMouseEnter={() => setIsVisible(true)}
+        onMouseLeave={() => setIsVisible(false)}
+        onFocus={() => setIsVisible(true)}
+        onBlur={() => setIsVisible(false)}
+      >
+        i
+      </button>
+      {isVisible && (
+        <div className="absolute left-1/2 -translate-x-1/2 bottom-full mb-2 w-48 p-2 bg-[#333333] text-white rounded-lg shadow-lg z-10" style={{ fontSize: '0.75em' }}>
+          {text}
+          <div className="absolute left-1/2 -translate-x-1/2 top-full w-0 h-0 border-l-4 border-r-4 border-t-4 border-transparent border-t-[#333333]" />
+        </div>
+      )}
+    </div>
+  );
+}
